Add TodayTodoArea.tsx and MonthlyTodoArea.tsx and Delete src/templates/MonthlyTodo.tsx
