+++ src/components/todo/page/monthly-todo-area/monthly-todo-area.scss
@@ -1,0 +1,6 @@
+
+.monthly-todo-area {
+  &__spacer {
+    margin-bottom: 20px;
+  }
+}
