+++ src/components/todo/page/todo-page.scss
@@ -46,7 +46,7 @@
     margin: 0 0 10px 10px;
   }
 
-  &__today-switch-todo-list {
+  &__switch-schedule-todo-list {
     background-color: #F4F5F7;
     border-radius: 0.6rem;
     padding: 25px 0;
