+++ src/components/uikit/tabs/switch-item-tabs.scss
@@ -1,8 +1,8 @@
 @import "../../../assets/variable";
 
-.switch-schedule-todo-list {
+.switch-item-tabs {
 
-  &__switch-button {
+  &__buttons {
     display: flex;
     justify-content: center;
     background-color: white;
