+++ src/reducks/groupTodoList/operations.ts
@@ -96,12 +96,14 @@ export const createGroupTodoListItem = (
         prevTodoList: GroupTodoList,
         nextTodoListItem: GroupTodoListItem
       ) => {
+        let nextTodoList: GroupTodoList = [];
         if (idx !== -1) {
           prevTodoList.splice(idx, 0, nextTodoListItem);
+          nextTodoList = prevTodoList.concat();
         } else if (idx === -1) {
-          prevTodoList.push(nextTodoListItem);
+          nextTodoList = prevTodoList.concat(nextTodoListItem);
         }
-        return prevTodoList;
+        return nextTodoList;
       };
 
       const groupExpiredTodoList = (prevTodoList: GroupTodoList, responseDate: string) => {
@@ -120,9 +122,9 @@ export const createGroupTodoListItem = (
       const groupTodayTodoList = (prevTodoList: GroupTodoList, responseDate: string) => {
         let nextTodoList: GroupTodoList = [];
         if (dateToDateString(today) === responseDate) {
-          nextTodoList = [newTodoListItem, ...prevTodoList];
+          nextTodoList = [newTodoListItem].concat(prevTodoList);
         } else if (dateToDateString(today) !== responseDate) {
-          return prevTodoList;
+          nextTodoList = prevTodoList.concat();
         }
         return nextTodoList;
       };
@@ -258,12 +260,14 @@ export const editGroupTodoListItem = (
         prevTodoList: GroupTodoList,
         nextTodoListItem: GroupTodoListItem
       ) => {
+        let nextTodoList: GroupTodoList = [];
         if (idx !== -1) {
           prevTodoList.splice(idx, 0, nextTodoListItem);
+          nextTodoList = prevTodoList.concat();
         } else if (idx === -1) {
-          prevTodoList.push(nextTodoListItem);
+          nextTodoList = prevTodoList.concat(nextTodoListItem);
         }
-        return prevTodoList;
+        return nextTodoList;
       };
 
       const updateGroupExpiredTodoList = (prevTodoList: GroupTodoList, responseDate: string) => {
@@ -276,7 +280,7 @@ export const editGroupTodoListItem = (
         if (dateToDateString(today) > responseDate) {
           if (result.data.complete_flag === true) {
             prevTodoList.splice(prevItemIdx, 1);
-            nextTodoList = [...prevTodoList];
+            nextTodoList = prevTodoList.concat();
           } else if (result.data.complete_flag === false) {
             if (prevItemIdx !== -1) {
               prevTodoList.splice(prevItemIdx, 1);
@@ -289,7 +293,7 @@ export const editGroupTodoListItem = (
             nextTodoList = pushResponseTodoListItem(idx, prevTodoList, result.data);
           }
         } else if (dateToDateString(today) <= responseDate) {
-          nextTodoList = [...prevTodoList];
+          nextTodoList = prevTodoList.concat();
         }
         return nextTodoList;
       };
@@ -301,10 +305,22 @@ export const editGroupTodoListItem = (
         );
 
         if (dateToDateString(today) === responseDate) {
-          prevTodoList[prevItemIdx] = result.data;
-          nextTodoList = [...prevTodoList];
+          if (prevItemIdx === -1) {
+            nextTodoList = [result.data].concat(prevTodoList);
+            return nextTodoList;
+          }
+
+          const prevCompleteFlag = prevTodoList[prevItemIdx].complete_flag;
+          if (prevCompleteFlag === result.data.complete_flag) {
+            prevTodoList.splice(prevItemIdx, 1);
+            nextTodoList = [result.data].concat(prevTodoList);
+          } else if (prevCompleteFlag !== result.data.complete_flag) {
+            prevTodoList[prevItemIdx] = result.data;
+            nextTodoList = prevTodoList.concat();
+          }
         } else if (dateToDateString(today) !== responseDate) {
-          nextTodoList = [...prevTodoList];
+          prevTodoList.splice(prevItemIdx, 1);
+          nextTodoList = prevTodoList.concat();
         }
         return nextTodoList;
       };
@@ -322,9 +338,10 @@ export const editGroupTodoListItem = (
 
         if (dateToYearAndMonthString(selectedDate) === responseMonth) {
           const prevCompleteFlag = prevTodoList[prevItemIdx].complete_flag;
+
           if (prevCompleteFlag !== result.data.complete_flag) {
             prevTodoList[prevItemIdx] = result.data;
-            nextTodoList = [...prevTodoList];
+            nextTodoList = prevTodoList.concat();
           } else if (prevCompleteFlag === result.data.complete_flag) {
             prevTodoList.splice(prevItemIdx, 1);
 
@@ -340,7 +357,7 @@ export const editGroupTodoListItem = (
           }
         } else if (dateToYearAndMonthString(selectedDate) !== responseMonth) {
           prevTodoList.splice(prevItemIdx, 1);
-          nextTodoList = [...prevTodoList];
+          nextTodoList = prevTodoList.concat();
         }
         return nextTodoList;
       };
